Move some references to 'models'
