+++ pkg/api/aim/controller/controller.go
@@ -6,13 +6,11 @@ import (
 
 // Controller handles all the input HTTP requests.
 type Controller struct {
-	appService        *AppService
+	appService        *service.AppService
 }
 
 // NewController creates new Controller instance.
-func NewController(
-	appService *AppService,
-) *Controller {
+func NewController(appService *service.AppService) *Controller {
 	return &Controller{
 		appService:        appService,
 	}
