+++ pkg/api/mlflow/api/response/experiment.go
@@ -9,7 +9,7 @@ import (
 	"github.com/rotisserie/eris"
 
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
+	"github.com/G-Research/fasttrackml/pkg/models"
 )
 
 // ExperimentTagPartialResponse is a partial response object for different responses.
