+++ pkg/api/mlflow/api/response/metric_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
+	"github.com/G-Research/fasttrackml/pkg/models"
 )
 
 func TestNewMetricHistoryResponse_Ok(t *testing.T) {
