+++ pkg/api/mlflow/dao/convertors/run.go
@@ -6,7 +6,7 @@ import (
 	"database/sql"
 
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
+	"github.com/G-Research/fasttrackml/pkg/models"
 	"github.com/G-Research/fasttrackml/pkg/database"
 )
