+++ pkg/api/mlflow/dao/convertors/run_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
+	"github.com/G-Research/fasttrackml/pkg/models"
 )
 
 func TestConvertCreateRunRequestToDBModel(t *testing.T) {
