+++ pkg/api/mlflow/dao/convertors/tag.go
@@ -2,7 +2,7 @@ package convertors
 
 import (
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
+	"github.com/G-Research/fasttrackml/pkg/models"
 )
 
 // ConvertSetRunTagRequestToDBModel converts request.SetRunTagRequest into actual models.Tag model.
